Fix for php cookie domain.
This fixes an issue that was caused if you were using a subdomain as
PlexAuth but a different subdomain for everything else. eg;
PlexAuth - secure.domain.com
sickrage - domain.com/sickrage

This is because the cookie created at secure.domain.com couldn't be
accessed as by default it is created with the domain restriction of the
current domain.